Break service dependency loop for kernel modules
Modules are not provided in images.

modules service is called before root is remounted read-write but we
needed a place to write them. Problem manifested in fsck, root and
sync-metadata service failing to start in the boot runlevel.

The service was split in a piece that can run without a read-write / by
storing modules in ramfs. This eats some memory early in the boot
process but modules amount to ~100MB while we have 2GB available so it
should be fine for the foreseeable future.

The second part runs later in the boot runlevel to synchronize modules
in /run to / and purge them from ramfs.